requirements.txt
pydantic>=2.0.0
pydantic-ai[logfire]
python-dotenv>=1.0.0
requests>=2.31.0
logfire>=1.0.0
google-cloud-vision>=3.4.4
google-cloud-storage>=2.10.0
portkey-ai>=0.2.0
pytest>=7.4.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0

